authentification

+++ backend/controllers/goal.controller.js
@@ -1,13 +1,13 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goal.model')
-
+const User = require('../models/user.model')
 
 // @desc    Get Goals
 // @route GET /api/goals
 // @access Private
 const getGoals =  asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({user: req.user.id})
     res.status(200).json(goals)
 })
 
@@ -20,7 +20,10 @@ const setGoal = asyncHandler(async (req, res) => {
         throw new Error('Pls add a text field')
     }
 
-    const goal = await Goal.create({text: req.body.text})
+    const goal = await Goal.create({
+        text: req.body.text,
+        user: req.user.id,
+    })
     res.status(200).json(goal)
 })
 
@@ -35,6 +38,17 @@ const updateGoal = asyncHandler(async (req, res) => {
         throw new Error('Goal not found')
     }
 
+    const user = await User.findById(req.user.id)
+    // check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+    // make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('user not authorized!')
+    }
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new:true})
     res.status(200).json(updatedGoal)
 })
@@ -48,6 +62,18 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if(!goal){
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+    // check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+    // make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+        res.status(401)
+        throw new Error('user not authorized!')
     }
 
     await goal.deleteOne()
